API & route refactoring and moved Ask Me threads to their own models

+++ server/mongodb/actions/Comment.js
@@ -1,5 +1,6 @@
 import mongoDB from "../index";
 import Comment from "../models/Comment";
+import AskMeThread from "../models/AskMeThread";
 import Thread from "../models/Thread";
 import User from "../models/User";
 import { getUser } from "./User";
@@ -20,22 +21,21 @@ export async function createComment(
     if (!user) {
       return Promise.reject(new Error("Invalid User ID for poster"));
     }
-    return Thread.findById(parentId).then(thread => {
-      if (!thread) {
-        return Promise.reject(new Error("Invalid Thread ID"));
-      }
 
-      return Comment.create({
-        poster,
-        parentId,
-        content,
-        postedAt,
-      });
+    return Comment.create({
+      poster,
+      parentId,
+      content,
+      postedAt,
     });
   });
 }
 
 export async function deleteComment(commentId) {
+  if (commentId == null) {
+    throw new Error("All parameters must be provided!");
+  }
+
   await mongoDB();
 
   return Comment.findOneAndDelete({ _id: commentId }).then(deletedComment => {
@@ -49,7 +49,44 @@ export async function deleteComment(commentId) {
   });
 }
 
+export async function getCommentsByAskMeThread(threadId) {
+  if (threadId == null) {
+    throw new Error("All parameters must be provided!");
+  }
+
+  await mongoDB();
+
+  return AskMeThread.findById(threadId)
+    .then(thread => {
+      if (!thread) {
+        return Promise.reject(new Error("Invalid Thread ID"));
+      }
+
+      return Comment.find({ parentId: threadId })
+        .sort({
+          postedAt: -1,
+        })
+        .then(async comments => {
+          if (!comments) {
+            return Promise.reject(new Error("Error retrieving comments"));
+          }
+
+          return Promise.all(
+            comments.map(async comment => ({
+              ...comment.toObject(),
+              author: await getUser(comment.poster),
+            }))
+          );
+        });
+    })
+    .catch(() => Promise.reject(new Error("Invalid Thread ID")));
+}
+
 export async function getCommentsByThread(threadId) {
+  if (threadId == null) {
+    throw new Error("All parameters must be provided!");
+  }
+
   await mongoDB();
 
   return Thread.findById(threadId)
@@ -58,18 +95,22 @@ export async function getCommentsByThread(threadId) {
         return Promise.reject(new Error("Invalid Thread ID"));
       }
 
-      return Comment.find({ parentId: threadId }).then(async comments => {
-        if (!comments) {
-          return Promise.reject(new Error("Error retrieving comments"));
-        }
+      return Comment.find({ parentId: threadId })
+        .sort({
+          postedAt: -1,
+        })
+        .then(async comments => {
+          if (!comments) {
+            return Promise.reject(new Error("Error retrieving comments"));
+          }
 
-        return Promise.all(
-          comments.map(async comment => ({
-            ...comment.toObject(),
-            author: await getUser(comment.poster),
-          }))
-        );
-      });
+          return Promise.all(
+            comments.map(async comment => ({
+              ...comment.toObject(),
+              author: await getUser(comment.poster),
+            }))
+          );
+        });
     })
     .catch(() => Promise.reject(new Error("Invalid Thread ID")));
 }
